Add upload dialog fragment
